requirements.txt
scikit-learn>=1.5.0
pandas>=2.2.2
numpy>=2.0.0
imbalanced-learn>=0.12.3
mlflow>=2.13.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.1
python-multipart>=0.0.9
slowapi>=0.1.9
python-jose[cryptography]>=3.3.0
streamlit>=1.35.0
plotly>=5.22.0
evidently>=0.4.30
pyyaml>=6.0.1
python-dotenv>=1.0.1
joblib>=1.4.2
requests>=2.32.2
dvc>=3.51.1
xgboost>=2.0.0
lightgbm>=4.0.0